+++ packages/services/server/src/index.ts
@@ -1,15 +1,9 @@
 #!/usr/bin/env node
 import got from 'got';
 import { GraphQLError, stripIgnoredCharacters } from 'graphql';
-import supertokens from 'supertokens-node';
-import {
-  errorHandler as supertokensErrorHandler,
-  plugin as supertokensFastifyPlugin,
-} from 'supertokens-node/framework/fastify/index.js';
 import cors from '@fastify/cors';
 import type { FastifyCorsOptionsDelegateCallback } from '@fastify/cors';
 import 'reflect-metadata';
-import { z } from 'zod';
 import formDataPlugin from '@fastify/formbody';
 import {
   createRegistry,
@@ -60,7 +54,6 @@ import { graphqlHandler } from './graphql-handler';
 import { clickHouseElapsedDuration, clickHouseReadDuration } from './metrics';
 import { createOtelAuthEndpoint } from './otel-auth-endpoint';
 import { createPublicGraphQLHandler } from './public-graphql-handler';
-import { initSupertokens, oidcIdLookup } from './supertokens';
 import { registerSupertokensAtHome } from './supertokens-at-home';
 
 class CorsError extends Error {
@@ -131,12 +124,6 @@ export async function main() {
       return res.status(403).send(err.message);
     }
 
-    if (env.supertokens.type === 'core') {
-      // We can not upgrade Supertokens Node as it removed some APIs we rely on for
-      // our SSO flow. This the as `any` cast here.
-      // The code is still compatible and purely a type error.
-      return supertokensErrorHandler()(err, req, res as any);
-    }
     server.log.error(err);
     return res.status(500);
   });
@@ -166,9 +153,11 @@ export async function main() {
           'graphql-client-name',
           'ignore-session',
           'x-request-id',
-          ...(env.supertokens.type === 'atHome'
-            ? ['rid', 'fdi-version', 'anti-csrf', 'authorization', 'st-auth-mode']
-            : supertokens.getAllCORSHeaders()),
+          'rid',
+          'fdi-version',
+          'anti-csrf',
+          'authorization',
+          'st-auth-mode',
         ],
       });
     };
@@ -396,10 +385,7 @@ export async function main() {
               emailVerification: env.auth.requireEmailVerification
                 ? registry.injector.get(EmailVerification)
                 : null,
-              accessTokenKey:
-                env.supertokens.type === 'atHome'
-                  ? new AccessTokenKeyContainer(env.supertokens.secrets.accessTokenKey)
-                  : null,
+              accessTokenKey: new AccessTokenKeyContainer(env.supertokens.secrets.accessTokenKey),
               oidcIntegrationStore: new OIDCIntegrationStore(storage.pool, redis, logger),
             }),
           organizationAccessTokenStrategy,
@@ -458,21 +444,7 @@ export async function main() {
       });
     }
 
-    if (env.supertokens.type == 'core') {
-      initSupertokens({
-        storage,
-        crypto,
-        logger: server.log,
-        redis,
-        taskScheduler,
-        broadcastLog,
-      });
-    }
-
     await server.register(formDataPlugin);
-    if (env.supertokens.type == 'core') {
-      await server.register(supertokensFastifyPlugin);
-    }
 
     await registerTRPC(server, {
       router: internalApiRouter,
@@ -529,42 +501,6 @@ export async function main() {
       },
     });
 
-    const oidcIdLookupSchema = z.object({
-      slug: z.string({
-        required_error: 'Slug is required',
-      }),
-    });
-
-    server.post('/auth-api/oidc-id-lookup', async (req, res) => {
-      const inputResult = oidcIdLookupSchema.safeParse(req.body);
-
-      if (!inputResult.success) {
-        captureException(inputResult.error, {
-          extra: {
-            path: '/auth-api/oidc-id-lookup',
-            body: req.body,
-          },
-        });
-        void res.status(400).send({
-          ok: false,
-          title: 'Invalid input',
-          description: 'Failed to resolve SSO information due to invalid input.',
-          status: 400,
-        } satisfies Awaited<ReturnType<typeof oidcIdLookup>>);
-        return;
-      }
-
-      const result = await oidcIdLookup(inputResult.data.slug, storage, req.log);
-
-      if (result.ok) {
-        void res.status(200).send(result);
-        return;
-      }
-
-      void res.status(result.status).send(result);
-      return;
-    });
-
     createOtelAuthEndpoint({
       server,
       authN,
@@ -581,18 +517,16 @@ export async function main() {
       return;
     });
 
-    if (env.supertokens.type === 'atHome') {
-      await registerSupertokensAtHome(
-        server,
-        storage,
-        registry.injector.get(TaskScheduler),
-        registry.injector.get(CryptoProvider),
-        registry.injector.get(RedisRateLimiter),
-        registry.injector.get(OAuthCache),
-        broadcastLog,
-        env.supertokens.secrets,
-      );
-    }
+    await registerSupertokensAtHome(
+      server,
+      storage,
+      registry.injector.get(TaskScheduler),
+      registry.injector.get(CryptoProvider),
+      registry.injector.get(RedisRateLimiter),
+      registry.injector.get(OAuthCache),
+      broadcastLog,
+      env.supertokens.secrets,
+    );
 
     if (env.cdn.providers.api !== null) {
       const s3 = {