+++ packages/web/app/src/pages/target-laboratory-new.tsx
@@ -26,9 +26,9 @@ import { graphql, useFragment } from '@/gql';
 import { TargetEnvPlugin } from '@/laboratory/plugins/target-env';
 import { useRedirect } from '@/lib/access/common';
 import { useLocalStorage, useToggle } from '@/lib/hooks';
-import { useCurrentOperation } from '@/lib/hooks/laboratory/use-current-operation';
+import { useCurrentOperationWithFetchingState } from '@/lib/hooks/laboratory/use-current-operation';
 import { TargetLaboratoryPageQuery } from '@/lib/hooks/laboratory/use-operation-collections-plugin';
-import { useSyncOperationState } from '@/lib/hooks/laboratory/use-sync-operation-state';
+import { useOperationFromQueryString } from '@/lib/hooks/laboratory/useOperationFromQueryString';
 import { useResetState } from '@/lib/hooks/use-reset-state';
 import { cn } from '@/lib/utils';
 import {
@@ -308,7 +308,7 @@ function useLaboratoryState(props: {
   targetSlug: string;
   defaultEndpoint: string | null;
 }) {
-  const [{ data, fetching }] = useQuery({
+  const [{ data, fetching: dataFetching }] = useQuery({
     query: LaboratoryQuery,
     variables: {
       selector: {
@@ -402,7 +402,7 @@ function useLaboratoryState(props: {
 
   const deleteOperation = useMemo(
     () =>
-      throttle((collection: LaboratoryCollection, operation: LaboratoryCollectionOperation) => {
+      throttle((_collection: LaboratoryCollection, operation: LaboratoryCollectionOperation) => {
         void mutateDelete({
           selector: {
             targetSlug: props.targetSlug,
@@ -470,17 +470,12 @@ function useLaboratoryState(props: {
     [mutateUpdatePreflight, props.targetSlug, props.organizationSlug, props.projectSlug],
   );
 
-  const currentOperation = useCurrentOperation({
-    organizationSlug: props.organizationSlug,
-    projectSlug: props.projectSlug,
-    targetSlug: props.targetSlug,
-  });
-
-  const { savedOperation } = useSyncOperationState({
-    organizationSlug: props.organizationSlug,
-    projectSlug: props.projectSlug,
-    targetSlug: props.targetSlug,
-  });
+  const { currentOperation, fetching: currentOperationFetching } =
+    useCurrentOperationWithFetchingState({
+      organizationSlug: props.organizationSlug,
+      projectSlug: props.projectSlug,
+      targetSlug: props.targetSlug,
+    });
 
   const router = useRouter();
   const { search } = router.latestLocation;
@@ -518,14 +513,14 @@ function useLaboratoryState(props: {
       return [...getLocalStorageState('operations', []), operationFromQueryString];
     }
 
-    if (currentOperation && savedOperation) {
+    if (currentOperation) {
       return [
         ...getLocalStorageState('operations', []),
         {
           id: currentOperation.id,
           name: currentOperation.name,
-          query: savedOperation.query,
-          variables: savedOperation.variables ?? '{}',
+          query: currentOperation.query,
+          variables: currentOperation.variables ?? '{}',
           headers: currentOperation.headers ?? '{}',
           extensions: '{}',
         } satisfies LaboratoryOperation,
@@ -533,7 +528,7 @@ function useLaboratoryState(props: {
     }
 
     return getLocalStorageState('operations', []);
-  }, [currentOperation, savedOperation, operationFromQueryString]);
+  }, [currentOperation, operationFromQueryString]);
 
   const defaultTabs = useMemo(() => {
     if (operationFromQueryString) {
@@ -547,7 +542,7 @@ function useLaboratoryState(props: {
       ];
     }
 
-    if (currentOperation && savedOperation) {
+    if (currentOperation) {
       return [
         ...getLocalStorageState('tabs', []),
         {
@@ -566,13 +561,17 @@ function useLaboratoryState(props: {
     }
 
     return getLocalStorageState('tabs', []);
-  }, [currentOperation, savedOperation, operationFromQueryString]);
+  }, [currentOperation, operationFromQueryString]);
 
-  // useEffect(() => {
-  //   if (search.operation) {
-  //     window.history.pushState(null, '', `${location.pathname}`);
-  //   }
-  // }, [search.operation]);
+  const operationIdFromSearch = useOperationFromQueryString();
+
+  const fetching = useMemo(() => {
+    if (operationIdFromSearch) {
+      return dataFetching || currentOperationFetching;
+    }
+
+    return dataFetching;
+  }, [dataFetching, currentOperationFetching, operationIdFromSearch]);
 
   return {
     fetching,