+++ packages/libraries/laboratory/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-hive/laboratory
 
+## 0.1.1
+
+### Patch Changes
+
+- [#7839](https://github.com/graphql-hive/console/pull/7839)
+  [`9585f11`](https://github.com/graphql-hive/console/commit/9585f11ef26db9ec9a3cb92d3758d1cf9e5655b7)
+  Thanks [@mskorokhodov](https://github.com/mskorokhodov)! - exposed renderLaboratory function to
+  use window location as default endpoint
+
 ## 0.1.0
 
 ### Minor Changes